app/requirements.txt
Django==3.2
djangorestframework
psycopg2-binary
gunicorn